Started Next Round Implemenation

+++ Assets/Scripts/Constants.cs
@@ -10,8 +10,8 @@ namespace Assets.Scripts
     {
         public const float turnPlayerDelay = 2.0f;
         public const float clearMethodDelay = 1.5f;
-        public const Int16 totalrounds = 5;
-        public const Int16 totalmatch = 5;
+        public const int totalrounds = 5;
+        public const int totalmatch = 5;
         
     }
 }
